+++ public/pages/Filters/components/FiltersTab.tsx
@@ -232,9 +232,7 @@ export const FiltersTab: React.FC<FiltersTabProps> = ({ spaceFilter, notificatio
       {itemForAction?.action === DELETE_ACTION && (
         <EuiConfirmModal
           title="Delete filter?"
-          onCancel={() => {
-            if (!isDeleting) setItemForAction(null);
-          }}
+          onCancel={() => setItemForAction(null)}
           onConfirm={confirmDeleteSingle}
           cancelButtonText="Cancel"
           confirmButtonText="Delete"
@@ -249,9 +247,7 @@ export const FiltersTab: React.FC<FiltersTabProps> = ({ spaceFilter, notificatio
       {itemForAction?.action === DELETE_SELECTED_ACTION && (
         <EuiConfirmModal
           title={`Delete ${selectedItems.length} ${pluralize(selectedItems.length, 'filter')}?`}
-          onCancel={() => {
-            if (!isDeleting) setItemForAction(null);
-          }}
+          onCancel={() => setItemForAction(null)}
           onConfirm={() => confirmDeleteSelected(selectedItems, () => setSelectedItems([]))}
           cancelButtonText="Cancel"
           confirmButtonText="Delete"