+++ public/pages/KVDBs/containers/KVDBs.tsx
@@ -263,9 +263,7 @@ export const KVDBs: React.FC<KVDBsProps> = ({ history, notifications }) => {
       {itemForAction?.action === DELETE_ACTION && (
         <EuiConfirmModal
           title="Delete KVDB"
-          onCancel={() => {
-            if (!isDeleting) setItemForAction(null);
-          }}
+          onCancel={() => setItemForAction(null)}
           onConfirm={confirmDeleteSingle}
           cancelButtonText="Cancel"
           confirmButtonText="Delete"
@@ -279,9 +277,7 @@ export const KVDBs: React.FC<KVDBsProps> = ({ history, notifications }) => {
       {itemForAction?.action === DELETE_SELECTED_ACTION && (
         <EuiConfirmModal
           title={`Delete ${selectedItems.length} KVDB${selectedItems.length !== 1 ? 's' : ''}`}
-          onCancel={() => {
-            if (!isDeleting) setItemForAction(null);
-          }}
+          onCancel={() => setItemForAction(null)}
           onConfirm={() => confirmDeleteSelected(selectedItems, () => setSelectedItems([]))}
           cancelButtonText="Cancel"
           confirmButtonText="Delete"